+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- using `SlideController` to control actions, instead of a standalone `ActionController`, [Issue](https://github.com/SimonWang9610/flutter_slidable_panel/issues/1)
+- [break change]: no longer using `SlideActionPanel`, users can create `SlidablePanel` more conveniently
+- stable
+
 ## 1.0.0
 
 - first version